requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
PyYAML==6.0.1